+++ Wall.cpp
@@ -1,7 +1,7 @@
 #include "Wall.h"
 
-Wall::Wall(int x, int y)
-	:SFMLObject(x, y,"Wall")
+Wall::Wall(float x, float y)
+	:SFMLObject(x, y, "Wall")
 {
 
 }
